Update LICENSE.md
chore: Replace modified MIT License with custom license

Removed the modified MIT License and created a new custom license due to the fact that the MIT License cannot be altered. The custom license is based on MIT but includes additional conditions:

1. Free Distribution: The Software cannot be sold. It must be distributed free of charge, without any monetary value associated.
2. Credit Maintenance: All copies or substantial portions of the Software must include the complete credits of the original authors and this unmodified, original license.
3. Inclusion of Copyright Notice and Permission: The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

This change ensures compliance with licensing norms and maintains the intended conditions for software distribution and credit.